fix(tests): Fix sign-conversion errors in Part 2 conformity tests for GCC compatibility
Resolves GitHub Actions CI build failure at 86% (Part 2 conformity tests).

PROBLEM:
- GCC with -Werror=sign-conversion treats bitwise NOT on hex literals as error
- Bitwise NOT on 0x01 produces int (-2 in two's complement), not uint8_t
- Assignment to unsigned char triggers signed->unsigned conversion warning

SOLUTION:
- Apply explicit static_cast<uint8_t> to all bitwise NOT operations
- Affected methods: set_professional(), set_linear_pcm(), set_lock_status()
- Lines 50, 55, 65: bytes[0] &= static_cast<uint8_t>(~0xNN)

VERIFICATION:
- Windows MSVC build: PASSING (no warnings)
- Part 2 tests: ALL 39/39 PASSING (10ms execution time)
- Semantics unchanged: explicit cast maintains correctness

IMPACT:
- Fixes Linux GCC CI build with strict type checking
- Maintains cross-platform compatibility (Windows + Linux)
- No functional change to test behavior

TRACES TO:
- GitHub Actions CI error: test_aes3_part2_conformity.cpp:50,55,65
- REQ-QUAL-META-001: AES3-2-2009 Part 2 conformity testing
- AES3-2-2009 Clause 5.5.0: Channel status byte 0 manipulation

+++ 05-implementation/tests/AES/AES3/Part2/_2009/conformity/test_aes3_part2_conformity.cpp
@@ -47,12 +47,12 @@ struct ChannelStatusBlock {
     // Byte 0 accessors (AES3-2-2009 Clause 5.5.0)
     bool is_professional() const { return (bytes[0] & 0x01) != 0; }
     void set_professional(bool prof) { 
-        if (prof) bytes[0] |= 0x01; else bytes[0] &= ~0x01; 
+        if (prof) bytes[0] |= 0x01; else bytes[0] &= static_cast<uint8_t>(~0x01); 
     }
     
     bool is_linear_pcm() const { return (bytes[0] & 0x02) == 0; }
     void set_linear_pcm(bool pcm) {
-        if (!pcm) bytes[0] |= 0x02; else bytes[0] &= ~0x02;
+        if (!pcm) bytes[0] |= 0x02; else bytes[0] &= static_cast<uint8_t>(~0x02);
     }
     
     uint8_t get_pre_emphasis() const { return (bytes[0] >> 2) & 0x07; }
@@ -62,7 +62,7 @@ struct ChannelStatusBlock {
     
     bool is_locked() const { return (bytes[0] & 0x20) == 0; }
     void set_lock_status(bool locked) {
-        if (!locked) bytes[0] |= 0x20; else bytes[0] &= ~0x20;
+        if (!locked) bytes[0] |= 0x20; else bytes[0] &= static_cast<uint8_t>(~0x20);
     }
     
     uint8_t get_sampling_frequency() const { return (bytes[0] >> 6) & 0x03; }
